WIP: lokalne izmene pre rebase-a

+++ database/migrations/2025_11_22_171951_add_distance_and_location_to_tours_table.php
@@ -9,7 +9,7 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('tours', function (Blueprint $table) {
-            $table->float('distance', 8, 2)->nullable()->after('difficulty');
+            $table->decimal('distance', 8, 2)->nullable()->after('difficulty');
             $table->string('location')->nullable()->after('distance');
         });
     }
